Initialised CherryPy
Included some cherrypy functionality with comments that clarify what everything does. 

Server is now findable (which is needed for the redirect)

Next step is to relay the contents from Bob to Alice. We know the website Alice tries to access (IP of Bob), hence this sould not be too difficult.